test : ScheduleService 순수 테스트 제거 후 integrationTest패키지를 service패키지와 통합
